Found the Buyer problem!!

+++ E-Store-Project-Updated/Buyer.cpp
@@ -197,13 +197,13 @@ int const Buyer::getTotalCartValue()const
 	int res = 0;
 
 	for (int i = 0; i < this->getNumberOfItems(); i++)
-		res += this->m_cart[i]->getPrice();
+		res +=(int)this->m_cart[i]->getPrice();
 	for (int i = 0; i < this->getNumOfOrders(); i++)
 	{
 		if (!this->m_checkout_orders[i]->getPaid())
 		{
 			for (int j = 0; j < this->m_checkout_orders[i]->getNumOfProducts(); j++)
-				res += this->m_checkout_orders[i]->getProductsArr()[j]->getPrice();
+				res += (int)this->m_checkout_orders[i]->getProductsArr()[j]->getPrice();
 		}
 	}
 	return res;
